Allow for a custom user name in the timeline
This change lets you set `twitter-screen-name="alternate_username"` to set the timeline to be fetched on `alternate_username` instead of the original one, regardless of how the widget was setup in the settings on Twitter itself. 

Note you still need the widget ID to be a valid one for this to work.